requirements.txt
Flask-WTF
requests

